encapsulate all formatting and whitespace in TokenWrite
